pyproject.toml
[tool.poetry]
name = "chatbot-clutch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "chatbot_clutch"}]

[tool.poetry.dependencies]
python = "^3.11"
jupyter = "^1.0.0"
pandas = "^2.0.3"
pyarrow = "^13.0.0"
langchain = "^0.0.279"
gpt4all = "^1.0.9"
lxml = "^4.9.3"
faker = "^19.3.1"
jupyter-black = "^0.3.4"
openai = "^0.28.0"
psycopg2-binary = "^2.9.7"
langchain-experimental = "^0.0.14"
python-dotenv = "^1.0.0"
llama-index = "^0.8.21"
torch = "^2.0.1"
transformers = "^4.33.0"
black = "^23.7.0"
sentence-transformers = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
